Use GeomType enumeration in more contexts
This also extracts the SystemLayoutInput class into its own module.
