+++ armi/physics/neutronics/globalFlux/tests/test_globalFluxInterface.py
@@ -10,6 +10,7 @@ from armi import settings
 from armi.physics.neutronics.globalFlux import globalFluxInterface
 from armi.reactor.tests import test_reactors
 from armi.reactor.tests import test_blocks
+from armi.reactor import geometry
 from armi.tests import ISOAA_PATH
 from armi.nuclearDataIO.cccc import isotxs
 
@@ -22,7 +23,8 @@ class MockParams:
 
 class MockCore:
     def __init__(self):
-        self.geomType = "spiral"  # dummy, not actually supported.
+        # just pick a random geomType
+        self.geomType = geometry.GeomType.CARTESIAN
         self.symmetry = "full"
         self.p = MockParams()
 
@@ -81,7 +83,7 @@ class TestGlobalFluxOptions(unittest.TestCase):
         reactor = MockReactor()
         opts = globalFluxInterface.GlobalFluxOptions("neutronics-run")
         opts.fromReactor(reactor)
-        self.assertEqual(opts.geomType, "spiral")
+        self.assertEqual(opts.geomType, geometry.GeomType.CARTESIAN)
 
 
 class TestGlobalFluxInterface(unittest.TestCase):
